set signal task id on some rpc thread
update
